% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pive.R
\name{cv.LJIVE}
\alias{cv.LJIVE}
\title{Fit the IVs model and perform Cross-Validation for LJIVE}
\usage{
cv.LJIVE(Y, X, Z, k, lambdaSeq, K = 10, intercept = TRUE, normalize = TRUE)
}
\arguments{
\item{Y}{Response variable}

\item{X}{Exposure variable}

\item{Z}{Instrumental variables}

\item{k}{Parameter for jackknife IVs}

\item{K}{Number of folds for cross-validation}

\item{intercept}{Logical, indicating whether to include an intercept (default: TRUE).}

\item{normalize}{Logical, indicating whether to normalize the instruments (default: TRUE).}
}
\value{
An object of class "LJIVE" containing various outputs.
}
\description{
Fit the IVs model and perform Cross-Validation for LJIVE
}
