% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pive.R
\name{pive}
\alias{pive}
\title{Lasso-type instrumental variables methods for estimating causal effects and selecting invalid instruments.}
\usage{
pive(Y, X, Z, bootstrap = TRUE, B = 500, alpha = 0.05, K = 10)
}
\arguments{
\item{Y}{Response variable}

\item{X}{Exposure variable}

\item{Z}{Instrumental variables matrix}

\item{bootstrap}{Logical, whether to perform bootstrap sampling (default is TRUE)}

\item{B}{Number of bootstrap samples (default is 500)}

\item{alpha}{Significance level for confidence intervals (default is 0.05)}

\item{K}{K-fold cross-validation (default is 10-fold CV)}
}
\value{
Estimated results
}
\description{
"pive" function applies six methods to estimate coefficients and performs bootstrap sampling to provide estimates, standard errors, and confidence intervals.
If bootstrap is FALSE, it provides causal estimates of all methods and the number of invalid instrumental variables.
}
