% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pive.R
\name{predict.LJIVE}
\alias{predict.LJIVE}
\title{Predict the model results based on beta and delta parameters}
\usage{
\method{predict}{LJIVE}(object, lambda, type = c("coefficients", "instruments"), ...)
}
\arguments{
\item{object}{A fitted model object of class "LJIVE".}

\item{lambda}{Numeric vector of lambda values for which predictions are desired.}

\item{type}{Character string specifying the type of prediction:
\itemize{
\item "coefficients" for predicting coefficients.
\item "instruments" for predicting instruments.
}}

\item{...}{Additional arguments passed to methods.}
}
\description{
Predict the model results based on beta and delta parameters
}
